+++ config/routes.js
@@ -1,7 +1,8 @@
 var routes = [
   { path: '/', controller: 'home', name: 'index' },
   { path: '/about', controller: 'home', name: 'about' },
-  { path: '/signup', controller: 'authentication', name: 'signup', pre: ['login_required', 'post'] },
+  { path: '/signup', controller: 'authentication', name: 'signup', middleware: ['login_required', 'post'] },
+  { path: '/login', controller: 'authentication', name: 'signup', middleware: ['login_required', 'post'] },
 ];
 
 router.route('/about')