+++ router.js
@@ -2,11 +2,13 @@ var express = require('express');
 
 var routes = require('./config/routes');
 var homeController = require('./app/controllers/homeController');
+var authenticationController = require('./app/controllers/authenticationController');
 
 var login_required = require('./config/passport').isAuthenticated;
 
 var controllers = {
-  'home': homeController
+  'home': homeController,
+  'authentication': authenticationController
 };
 
 var pre = {
@@ -30,35 +32,36 @@ function routing_ops(app) {
       if (route.pre) {
 
         if (typeof route.pre === typeof 'String') {
-          controller_component.get && app.get(route.path, pre[route.pre], controller_component.get);
-          controller_component.post && app.post(route.path, pre[route.pre], controller_component.post);
-          controller_component.put && app.put(route.path, pre[route.pre], controller_component.put);
-          controller_component.delete && app.delete(route.path, pre[route.pre], controller_component.delete);
+          controller_component.get && app.get(route.path, pre[route.pre[0]], controller_component.get);
+          controller_component.post && app.post(route.path, pre[route.pre[0]], controller_component.post);
+          controller_component.put && app.put(route.path, pre[route.pre[0]], controller_component.put);
+          controller_component.delete && app.delete(route.path, pre[route.pre[0]], controller_component.delete);
         } else if (typeof route.pre === typeof ['Array']) {
 
           if (route.pre.indexOf('get') >= 0) {
-            controller_component.get && app.get(route.path, pre[route.pre], controller_component.get);
+            controller_component.get && app.get(route.path, pre[route.pre[0]], controller_component.get);
           } else {
             controller_component.get && app.get(route.path, controller_component.get);
           }
 
           if (route.pre.indexof('post') >= 0) {
-            controller_component.post && app.post(route.path, pre[route.pre], controller_component.post);
+            controller_component.post && app.post(route.path, pre[route.pre[0]], controller_component.post);
           } else {
             controller_component.post && app.post(route.path, controller_component.post);
           }
 
           if (route.pre.indexof('put') >= 0) {
-            controller_component.put && app.put(route.path, pre[route.pre], controller_component.put);
+            controller_component.put && app.put(route.path, pre[route.pre[0]], controller_component.put);
           } else {
             controller_component.put && app.put(route.path, controller_component.put);
           }
 
-          if (route.pre.indexof('put') >= 0) {
-            controller_component.delete && app.delete(route.path, pre[route.pre], controller_component.delete);
+          if (route.pre.indexof('delete') >= 0) {
+            controller_component.delete && app.delete(route.path, pre[route.pre[0]], controller_component.delete);
           } else {
             controller_component.delete && app.delete(route.path, controller_component.delete);
           }
+
         }
       } else {
         controller_component.get && app.get(route.path, controller_component.get);