+++ HISTORY.md
@@ -1,6 +1,7 @@
 # 0.6.0
 
 - Add more dependencies for authentication
+- Add User model
 
 # 0.5.0 / 2014-10-31